nbohr1more: more int to long conversions for 3954
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/x64@6662 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ framework/async/MsgChannel.h
@@ -69,8 +69,8 @@ private:
 	byte			ReadByte( void );
 	void			WriteShort( int s );
 	int				ReadShort( void );
-	void			WriteInt( int l );
-	int				ReadInt( void );
+	void			WriteLong( int l );
+	int				ReadLong( void );
 	void			WriteData( const byte *data, const int size );
 	void			ReadData( byte *data, const int size );
 };
